Menu backend server error

+++ src/main/java/com/ebarapp/ebar/repository/BillRepository.java
@@ -1,3 +1,4 @@
+
 package com.ebarapp.ebar.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -6,7 +7,6 @@ import org.springframework.stereotype.Repository;
 import com.ebarapp.ebar.model.Bill;
 
 @Repository
-public interface BillRepository extends JpaRepository<Bill, Long> {
+public interface BillRepository extends JpaRepository<Bill, Integer> {
 
 }
-
