+++ src/main/java/com/ebarapp/ebar/repository/ItemMenuRepository.java
@@ -1,3 +1,4 @@
+
 package com.ebarapp.ebar.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -6,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import com.ebarapp.ebar.model.ItemMenu;
 
 @Repository
-public interface ItemMenuRepository extends JpaRepository<ItemMenu, Long> {
+public interface ItemMenuRepository extends JpaRepository<ItemMenu, Integer> {
 
 }
